pyWordle/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.16.1
python-datamuse==2.0.1
requests==2.31.0
rich==13.5.2
urllib3==2.0.5
